+++ src/main/java/com/ssg/webpos/dto/ReissueTokenResponseDTO.java
@@ -1,0 +1,13 @@
+package com.ssg.webpos.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReissueTokenResponseDTO {
+    private String accessToken;
+    private String refreshToken;
+}
